Make _hashCode readonly
It's good practice to only use read-only fields in the GetHashCode()
method. Explicitly making the field `readonly` will help avoid bugs in
the future.

I also changed the `CreateHashCode()` method to return a value instead
of modifying the `_hashCode` field. This helps to solidify the notion
that the `MetadataCacheKey` class is immutable.