Protect against NullReferenceExceptions when Load() returns a null profiler
There is a race condition where getting the list of profiler GUIDs would return a valid id but then loading the profiler by id it might not exist. This causes NullRefereneceExceptions. For example, we have our own Redis storage implementation where profilers automatically expire after a specified timeout.

So, depending on the storage, profilers might not exist by id and this protects against that.