Fix bin directory
\bin is C:\Bin (oops), but I didn't notice this as it followed a
successful build in which the bin\ directory was fine, just decaying in
age.
Unfortuantely bin\ is not correct either, since <OutputPath> is being
overriden in the SDK (see https://github.com/dotnet/sdk/issues/550).
Currently, this is outputting to bin\net462\, which isn't where an MVC5
expects it. I'm blocked a bit on samples until I can get around this
issue.