fixes #173
There is an issue in ASP.NET when the profiler is started in BeginRequest, but
accessed in an HttpHandler later on, in which `Head` is null (it should not be
null until the profiler is stopped).

The `Head` property is now backed by `AsyncLocal<T>` which is tied to the current
thread's `ExecutionContext`. However the `ExecutionContext` does not flow between
ASP.NET pipeline events if the events fire on separate threads.

This fix (a hack really) ensures that `Head` is not null by setting it to `Root`.
In practice I think this will be OK since it's unlikely a user will create a timing,
other than root, which spans pipeline events.