Move to WebRequestProfilerProvier.Setup()
The previous way was confusing if you mis-ordered the RouteBasePath and
the creation of WebRequestProfilerProvier, since it registered the
routes to use. If you got this backwards it just silently failed, and a
sample will never make this crystal clear.

So instead, let's move that into a method where you can't mess it up -
that's far saner. Fixes #185