Massive refactor of options
This is a series of changes that aims to:
- Decouple several concepts and interfaces from one another
- Make Options the root object a user needs, not many things
- Remove all statics except those needed for MiniProfiler.Current
- Be more friendly to initial setup (on both platforms)
- Ease option configurations
- Have more sensible defaults
- Simpify EntityFramework integration
- Increase testability (be removing most statics)

There's more to do here, this is an initial commit for the next alpha. Next up:
- .AspNetCore tests
- Find a better name for WebRequestProfilerProvider (should be AspNetClassicProvider or something)
- Update all the docs for these changes
- Simplify storage spinup (and make IMemoryCache work without a user specifying anything using ASP.NET Core 2.x constructs)
- Sanity check for nulls with all this movement

This also upgrades xUnit to incldue my PR which allows the async tests to run at the end by themselves and be alot more stable on timings for fewer-core machines.