Tag helpers for client timing scripts (#304)
Two tag helpers to initialize client timing scripts and to wrap a section of code for timing.  The initialization tag is optional as it will be pre-pended to the timing script if it hasn't already been loaded.  Also, the initialization script will only render once in case someone accidentally adds it twice.

Code below is how I was testing the behavior for the helpers.

```
<mini-profiler-timing />
<mini-profiler-timing />
<div></div>
<profile name="Render JS delay">
	<profile-script name="JS delay">
		<script language="javascript">
			alert("Wait for a second");
		</script>
	</profile-script>
</profile>
```

This partially addresses issues I mentioned in issue #302.