Exposed End Points Added to Documentation (#386)
Addressing Issue #342

Looked at the ProcessRequest method in MiniProfilerHandler. I added it to the overview page as the routes don't changed based on framework and thus should be on a shared page. Further, creating a separate page seemed unwarranted for the amount of information.

Any other content that should be covered or any routes that I missed?