.NET Core 3: Add Server-Timing trailer
This adds an option (default off) to enable Server-Timing headers for requests. It's a trailer here so requires Kestrel/feature support. If this works well, we can look at adding a netcoreapp2.2 build for MiniProfiler.AspNetCore* which should also support the headers.

I think this works, but it's hard to verify because of an https:// issue in https://github.com/aspnet/AspNetCore/issues/8952 which breaks Chrome and Firefox that won't be resolved until preview 7. Edge can connect (due to never fixing their TLS leniency...), but not show the trailers. Figures :)