DARK MODE! (#451)
Made to address #450...and because it's fun! Since MiniProfiler no longer supports IE 11, we can take advantage of CSS variables and help some users here.

Note: this isn't _only_ dark mode. It's really 2 things:
- Convert all colors (and the rounded borders) to CSS variables
- Introduce a dark mode via selectors that override those variables

Overall, this also allows a user to theme these variables however they want - it's not limited to the 2 included themes, that's just what's "in-box". I need to consult some experts but I'd love to also add a "High Contrast" theme to help another set of users...we just need color codes for it.

Here are some screens of dark mode (light is equivalent to today):
![image](https://user-images.githubusercontent.com/454813/72737986-9cb88080-3b6e-11ea-8a44-243a85f1c5c2.png)
![image](https://user-images.githubusercontent.com/454813/72737994-9f1ada80-3b6e-11ea-89a7-59b805974e12.png)
![image](https://user-images.githubusercontent.com/454813/72737997-a0e49e00-3b6e-11ea-9bdd-b3ffe195b5b8.png)
![image](https://user-images.githubusercontent.com/454813/72738000-a2ae6180-3b6e-11ea-9d6c-b570cf6869de.png)

Full pages respect this theme setting as well:
![image](https://user-images.githubusercontent.com/454813/72749626-933c1200-3b88-11ea-8dd0-044d9428b2fa.png)
![image](https://user-images.githubusercontent.com/454813/72749628-9505d580-3b88-11ea-987b-579ce7f0d1b5.png)