Pretty up and don't allocate descriptor names
Makes action descriptor names a bit cleaner by removing the assembly name. Also caches so we only discard a cheaper tuple for lookup only. Could change to per-type methods and eliminate the tuple as well, or a dictionary based on key, or something else. Just depends how crazy we want to go here.