The the problem!
This sets the parent profiler to where it should be when adding a new timing, even in the case of stack unwinds further up which hide our understanding of a previous head reset further down due to an AsyncLocal<Timing> mismatch on the head. Now as a safety we roll up to the parent and get the right tree even with MVCs frame-saving shenanigans. Woohoo!

Good god that was hard to figure out.