More work on ASP.NET Core timings (#482)
* More work on ASP.NET Core timings

This is a stab, may change a lot...getting it up for tweaking. Trying to repro the negative timing issue we see in Stack Overflow.

* DEBUG MODE OF DOOM

Gotta figure out stack slicing off the top here yet...but some progress in an uber debug mode for things.

* More style and tooltip refinement

* Stack shaving and tooltip for trivials fixed

* The the problem!

This sets the parent profiler to where it should be when adding a new timing, even in the case of stack unwinds further up which hide our understanding of a previous head reset further down due to an AsyncLocal<Timing> mismatch on the head. Now as a safety we roll up to the parent and get the right tree even with MVCs frame-saving shenanigans. Woohoo!

Good god that was hard to figure out.

* Remove Debug overhead shenanigans.

* Fix debug off TypeScript

* Doc updates

* Undo some diff, add inlining hints

* Doc tidy

* Fix stack styling (undo removal here)

* Undo more diff

* More doc additions