Use Razor Pages route value in MiniProfiler name (#495)
If you are using Razor Pages instead of controllers and views MiniProfiler shows all result names in the results list as `/`.  This is because `routeData.Values["controller"]` and `routeData.Values["action"]` are null.  

This change checks if `routeData.Values["pages"]` has a value and uses that for the MiniProfiler name if available.

I've also added a Razor Page to the Asp.Net Core 3 sample for testing